add page for creating new tag name

+++ src/main/java/com/nppgks/reports/entity/ReportType.java
@@ -2,10 +2,7 @@ package com.nppgks.reports.entity;
 
 import lombok.Data;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
+import javax.persistence.*;
 
 @Entity(name = "report_type")
 @Data
@@ -13,14 +10,12 @@ public class ReportType {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String name;
 
     private String description;
 
     private int active;
-
-
 
 }
