+++ src/main/java/com/nppgks/reports/repository/ReportTypeRepository.java
@@ -3,6 +3,6 @@ package com.nppgks.reports.repository;
 import com.nppgks.reports.entity.ReportType;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportTypeRepository extends JpaRepository<ReportType, Long> {
+public interface ReportTypeRepository extends JpaRepository<ReportType, Integer> {
 
 }
